requirements.txt
pyautogui
pywin32==225
pypiwin32
comtypes==1.1.4
dragonfly2
keyboard
